melhorias visuais pagina login, outras...

+++ database/migrations/2016_07_19_160316_create_clientes_table.php
@@ -35,7 +35,7 @@ class CreateClientesTable extends Migration
             $table->integer('numero');
             $table->double('valorContrato');
             $table->integer('quantidadeMeses');
-            $table->integer('desconto');
+            $table->double('desconto');
 
             $table->string('email', 190)->unique();
